twitter_trend_view/create_empty_sheet/requirements.txt
# Function dependencies, for example:
# package>=version

pytz
openpyxl
google-cloud
google-cloud-storage
